add e2e case
