+++ tests/images/e2e/Dockerfile
@@ -2,6 +2,7 @@ FROM debian:buster-slim
 
 ENV KUBECTL_VERSION=v1.12.2
 ENV HELM_VERSION=v2.9.1
+ENV RELEASED_VERSION=v1.0.6
 
 RUN apt-get update && \
     apt-get install -y ca-certificates curl git openssl default-mysql-client unzip
@@ -28,6 +29,22 @@ ADD bin/e2e.test /usr/local/bin/
 ADD bin/webhook /usr/local/bin/
 ADD bin/blockwriter /usr/local/bin/
 ADD bin/apiserver /usr/local/bin/
+
+RUN mkdir /charts/${RELEASED_VERSION}
+
+RUN curl -L https://github.com/pingcap/tidb-operator/releases/download/${RELEASED_VERSION}/tidb-operator-chart-${RELEASED_VERSION}.tgz \
+    -o tidb-operator-chart-${RELEASED_VERSION}.tgz && \
+    tar -zxvf tidb-operator-chart-${RELEASED_VERSION}.tgz && \
+    mv tidb-operator /charts/${RELEASED_VERSION}/tidb-operator && \
+    rm -rf tidb-operator && \
+    rm tidb-operator-chart-${RELEASED_VERSION}.tgz
+
+RUN curl -L https://github.com/pingcap/tidb-operator/releases/download/${RELEASED_VERSION}/tidb-cluster-chart-${RELEASED_VERSION}.tgz \
+    -o tidb-cluster-chart-${RELEASED_VERSION}.tgz && \
+    tar -zxvf tidb-cluster-chart-${RELEASED_VERSION}.tgz && \
+    mv tidb-cluster /charts/${RELEASED_VERSION}/tidb-cluster && \
+    rm -rf tidb-cluster && \
+    rm tidb-cluster-chart-${RELEASED_VERSION}.tgz
 
 ADD entrypoint.sh /usr/local/bin
 ENTRYPOINT ["/usr/local/bin/entrypoint.sh"]
